Refactor parse source as prep for read timeouts
The read timeout crate requires the Read object to also be AsRawFd, so
it can use polling to check if there's anything to read. This broke
everything since `Box<Read + AsRawFd>` isn't valid. To fix this I had to
make my own source type that implemented both of these traits.