Collect all the lines of a log message
This goes back to the previous model of having 1 log message struct that
contains all the lines of a log message. It now works with tailing
because of read timeouts from the timeout-readwrite crate.

This enables the message in multi-line log messages to be started on the
next line so it aligns with the rest of the message lines.